% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipft.R
\name{ipfEstimateBeaconPositions}
\alias{ipfEstimateBeaconPositions}
\title{Estimates the positions of the emitter beacons}
\usage{
ipfEstimateBeaconPositions(fingerprints, positions, method = "wcentroid",
  rssirange = c(-100, 0), norssi = NA)
}
\arguments{
\item{fingerprints}{a data frame or a matrix with the RSSI fingerprints}

\item{positions}{a data frame or a matrix with the positions of the fingerprints}

\item{method}{method to use to estimate the position of the access points:
'centroid', 'wcentroid' or 'wip'}

\item{rssirange}{a numeric vector with the range of the RSSI data}

\item{norssi}{value used in dataRSSI when a beacon is not detected}
}
\description{
Estimates the positions of the emitter beacons
}
\examples{

    wapp <- ipfEstimateBeaconPositions(ipftrain[, 1:168], ipftrain[, 169:170], method = 'wcentroid')

}
