% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipft.R
\name{ipfGroup}
\alias{ipfGroup}
\title{Creates groups based on the specified parameters}
\usage{
ipfGroup(data, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Variables to group by. All variables (columns) will
be used if no parameter is provided.}
}
\value{
A numeric vector with the ids of the groups, in the
            same order as they appear in the data provided.
}
\description{
This function groups the data based on the specified variables
and assigns an id to each group
}
\examples{

    group <- ipfGroup(mtcars, cyl)

    group <- ipfGroup(mtcars, gear, carb)

    group <- ipfGroup(ipftrain, X, Y)

}
