% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipft.R
\name{ipfPlotLocation}
\alias{ipfPlotLocation}
\title{Plots the spatial location of the observations}
\usage{
ipfPlotLocation(positions, plabel = FALSE, reverseAxis = FALSE,
  xlab = NULL, ylab = NULL, title = "", pgrid = FALSE)
}
\arguments{
\item{positions}{a data frame or matrix with the positions}

\item{plabel}{if TRUE, adds labels to groups / observations}

\item{reverseAxis}{swaps axis}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{title}{plot title}

\item{pgrid}{plot grid (boolean)}
}
\description{
Plots the spatial location of the observations
}
\examples{

    ipfPlotLocation(ipftrain[, 169:170])

    ipfPlotLocation(ipftrain[, 169:170], plabel = TRUE, reverseAxis = TRUE,
                    title = 'Position of training set observations')

}
