% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipft.R
\name{ipfPlotPdf}
\alias{ipfPlotPdf}
\title{Plots the probability density function of the estimated error}
\usage{
ipfPlotPdf(estimation, xlab = "error", ylab = "density",
  title = "Probability density function")
}
\arguments{
\item{estimation}{an ipfEstimation}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{title}{plot title}
}
\description{
Plots the probability density function of the estimated error
}
\examples{

    model <- ipfKnn(ipftrain[, 1:168], ipftrain[, 169:170])
    estimation <- ipfEstimate(model, ipftest[, 1:168], ipftest[, 169:170])
    ipfPlotPdf(estimation)

}
