% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipfpwap}
\alias{ipfpwap}
\title{Indoor localization data set with the positions of the wireless access points present
in the ipftrain and ipftest data sets. Unknown locations are stored as NAs.
Data from the positioning tutorial of the seventh international conference on indoor Positioning and Indoor
Navigation (IPIN2016).}
\format{A data frame with columns:
\describe{
 \item{X}{X coordinate in meters relative to the origin of a predefined Cartesian coordinate system. From -61 to 50}
 \item{Y}{Y coordinate in meters relative to the origin of a predefined Cartesian coordinate system. From 8.5 to 117.5}
}}
\source{
UJI - Institute of New Imaging Technologies, Universitat Jaume I,
Avda. Vicente Sos Baynat S/N, 12071, Castellón, Spain. \url{http://www.init.uji.es/}
}
\usage{
ipfpwap
}
\description{
Indoor localization data set with the positions of the wireless access points present
in the ipftrain and ipftest data sets. Unknown locations are stored as NAs.
Data from the positioning tutorial of the seventh international conference on indoor Positioning and Indoor
Navigation (IPIN2016).
}
\examples{
\dontrun{
 ipfpwap
}
}
\keyword{datasets}
